test: keep fs sandbox tests inline
